feat: update dvpm to v5.1.2 and conditionally load telescope plugins
This commit updates the `@yukimemi/dvpm` dependency to version 5.1.2.  Additionally, it makes the loading of telescope plugins conditional, only enabling them if the `telescope` plugin status is true. This improves flexibility and allows for disabling telescope-related functionality when needed.  The `octo.nvim` plugin now depends on the `telescope` plugin status.  The `telescope` plugin itself is now conditionally loaded.