feat: change key mapping for Telescope frecency
The key mapping for triggering Telescope frecency is changed from `<c-p>` to `<space>fo`.
This change was made to avoid conflicts with other key mappings and to provide a more consistent and memorable key binding for accessing the frecency list.

The mapping for Telescope oldfiles in telescope.lua was also commented out, as the `<space>fo` mapping is now used for frecency.