feat: Enable render-markdown.nvim and simplify config
This commit enables the `render-markdown.nvim` plugin and simplifies its configuration.

The plugin's `enabled` property is set to `true` in `filetypes.ts`.
The configuration of the plugin is simplified in `render-markdown.lua`, removing the custom setup and using the default one instead.
Additionally, the keymaps for yazi were swapped to be more intuitive.