fix: Update dvpm dependency and enable avante plugin
This commit updates the dvpm dependency to version 6.2.2, and enables the avante plugin with build script.
The avante plugin is now enabled under the "full" profile. Additionally, a build step has been added to the avante plugin.
This step will execute a build script when the plugin is updated or loaded on windows or a make command if not on windows.