feat: enable ccc.nvim and increase spectrism interval
This commit enables the ccc.nvim plugin for "full" profiles and increases the spectrism interval and checkwait.

The ccc.nvim plugin provides color conversion and manipulation utilities.

The spectrism interval was increased from 60 seconds to 3600 seconds, and spectrism checkwait was increased from 1000ms to 3000ms, to reduce the frequency of color scheme updates.