feat: Add support for legacy Outlook
This commit adds support for activating the legacy version of Outlook.

The script now checks for the existence of a file `\.autohotkey\usenewoutlook` in the user's profile directory. If the file exists, the script activates the new Outlook app located in `\Microsoft\WindowsApps\olk.exe`. Otherwise, it activates the legacy Outlook application located in `C:\Program Files\Microsoft Office\root\Office16\OUTLOOK.EXE`.

This change allows users who are still using the legacy version of Outlook to continue using the `^F9` hotkey to activate the application.